secure-chat-app/requirements.txt
flask
flask-socketio
eventlet
cryptography
python-socketio
